requirements.txt
streamlit
tensorflow-cpu==2.13.0
opencv-python-headless
numpy